fixed problem when used with no domain

+++ CyberGetBadPasswords.ps1
@@ -8,7 +8,11 @@ $ACQ = ACQ("GetBadPasswords")
 $dnsroot =  (Get-ADDomain).dnsroot
 $DistinguishedName = (Get-ADDomain).DistinguishedName
 Write-Host "Changing default Domain config settings to $dnsroot and $DistinguishedName"
-Copy-Item "$GBPAppPath\Get-bADpasswords.ps1" "$GBPAppPath\Get-bADpasswords.ps1.bak"
+if (Test-Path -Path "$GBPAppPath\Get-bADpasswords.ps1.bak") {
+    Copy-Item "$GBPAppPath\Get-bADpasswords.ps1.bak" "$GBPAppPath\Get-bADpasswords.ps1"    
+}else {
+    Copy-Item "$GBPAppPath\Get-bADpasswords.ps1" "$GBPAppPath\Get-bADpasswords.ps1.bak"    
+}
 $domainConfig = Get-Content "$GBPAppPath\Get-bADpasswords.ps1"
 $newConfig = (($domainConfig -replace "YourDomainName", $dnsroot) -replace 'DC=domain,DC=com', $DistinguishedName) -replace "Send-MailMessage","#Send-MailMessage"
 $newConfig | Set-Content "$GBPAppPath\Get-bADpasswords.ps1"
